requirements.txt
langchain>=0.1.0
langgraph>=0.0.17
chromadb>=0.4.18
langchain-community>=0.0.13
langchain-chroma>=0.0.6
pydantic>=2.5.2
python-dotenv>=1.0.0
pypdf>=3.17.1
langchain-ollama>=0.0.1
flask>=2.0.0
werkzeug>=2.0.0 